Module Repr.Attribute

Attributes provide a mechanism for attaching metadata to type representations.

Attributes provide a mechanism for attaching metadata to type representations.

val set_random : (Stdlib.Random.State.t -> 'a) -> 'a Repr__.Type.t -> 'a Repr__.Type.t