Module Repr__.Attribute

type 'f t

An 'f t is an attribute key that can be used to pack polymorphic data into a heterogeneous Map (and then recover it again).

The type parameter 'f is the brand of a type operator f : * ⇒ * which, when applied to the type parameter 'a of a Map.t, gives the type 'a f of the associated data. This allows a single attribute key to store polymorphic data.

val create : name:string -> _ t

create ~name is a fresh attribute key with the given string name.

val name : _ t -> string

Get the string name of an attribute key.

module Map : sig ... end

module type S1 = Repr__.Attribute_intf.S1 with type 'a attr := 'a t and type 'a map := 'a Map.t

module Make1 : functor (T : sig ... end) -> S1 with type 'a t = 'a T.t