Module DSL.Size

type 'a t = private 'a Repr.Size.t =

| Static of int

| Dynamic of 'a

| Unknown

A value representing information known about the length in bytes of encodings produced by a particular binary codec:

Static n: all encodings produced by this codec have length n;

Dynamic f: the length of binary encodings is dependent on the specific value, but may be efficiently computed at run-time via the function f.

Unknown: this codec may produce encodings that cannot be efficiently pre-computed.

val of_value : 'a ty -> ('a -> int) t

val of_encoding : 'a ty -> (string -> int -> int) t

type -'a sizer = 'a size_of

val t : 'a ty -> 'a sizer

val using : ('b -> 'a) -> 'a sizer -> 'b sizer

val custom_static : int -> _ sizer

val custom_dynamic : ?⁠of_value:('a -> int) -> ?⁠of_encoding:(string -> int -> int) -> unit -> 'a sizer