Module Type_core.Encode_json

type 'a attr

type 'a map

type 'a t = 'a encode_json

The type of data associated with the attr attribute key.

val add : 'a t -> 'a map -> 'a map

Attach data for attr to a given map.

val find : 'a map -> 'a t option

Search for data corresponding to the key attr in the given map.

include Repr__.Higher.Branded.S with type 'a t := 'a t

type 'a t

type br

val inj : 'a t -> ('a, br) Repr__.Higher.app

val prj : ('a, br) Repr__.Higher.app -> 'a t

val key : br attr