Module Repr__Attribute_intf

module type S1 = sig ... end

An attribute key is a value that can be used to attach polymorphic data to a heterogeneous attribute map.

module type Attribute = sig ... end