Module Repr__Size.Sizer

type 'a t = {

of_value : ('a -> int) size;

of_encoding : (string -> Offset.t -> Offset.t) size;

}

An 'a t is a value that represents the size information known about a particular codec for type 'a.

of_value: given a value to encode, return the size of its encoding.

of_encoding: given a buffer buf and an offset off, return the _offset_ immediately _after_ the encoding starting at buf.[off] NOTE: not the length of the encoding itself, to enable chains of such sizers to call each other in tail-position.

Invariant: ∀ n. (of_value = Static n) ⟺ (of_encoding = Static n).

val (<+>) : a. 'a t -> 'a t -> 'a t

val static : int -> 'a t

val dynamic : of_value:('a -> int) -> of_encoding:(string -> Offset.t -> Offset.t) -> 'a t

val using : ('a -> 'b) -> 'b t -> 'a t

val unknown : 'a t